requirements.txt
streamlit
sentence-transformers
scikit-learn
pandas
numpy
matplotlib
fastapi 
uvicorn[standard]  
python-multipart
google-generativeai